requirements.txt
fastmcp>=2.12.0
uvicorn>=0.35.0
python-dotenv>=1.0.0
icalendar>=5.0.0
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.97.0
aiohttp>=3.8.5
